Extract archive rendering related controller code
In admin attachments and admin elements controllers we use the
same code to handle the archive rendering. Either in an overlay
dialog (if a content_id is available) or as normal html index view.
