+++ app/controllers/alchemy/admin/pictures_controller.rb
@@ -4,6 +4,7 @@ module Alchemy
   module Admin
     class PicturesController < Alchemy::Admin::ResourcesController
       include UploaderResponses
+      include ArchiveOverlay
 
       helper 'alchemy/admin/tags'
 
@@ -140,19 +141,6 @@ module Alchemy
         when 'large' then 12
         else
           20
-        end
-      end
-
-      def in_overlay?
-        params[:content_id].present?
-      end
-
-      def archive_overlay
-        @content = Content.find_by(id: params[:content_id])
-
-        respond_to do |format|
-          format.html { render partial: 'archive_overlay' }
-          format.js   { render action:  'archive_overlay' }
         end
       end
